+++ core/bin/dust_api.rs
@@ -22,10 +22,10 @@ use dust::{
     databases_store::store::{self as databases_store, DatabasesStore},
     dataset,
     deno::js_executor::JSExecutor,
-    filter::{Filterable, SearchFilter},
     project,
     providers::provider::{provider, ProviderID},
     run,
+    search_filter::{Filterable, SearchFilter},
     sqlite_workers::client::{self, HEARTBEAT_INTERVAL_MS},
     stores::{postgres, store},
     utils::{self, error_response, APIError, APIResponse, CoreRequestMakeSpan},
