+++ core/src/data_sources/data_source.rs
@@ -4,11 +4,11 @@ use crate::consts::DATA_SOURCE_DOCUMENT_SYSTEM_TAG_PREFIX;
 use crate::data_sources::qdrant::{QdrantClients, QdrantDataSourceConfig};
 use crate::data_sources::splitter::{splitter, SplitterID};
 use crate::databases_store::store::DatabasesStore;
-use crate::filter::{Filterable, SearchFilter};
 use crate::project::Project;
 use crate::providers::embedder::{EmbedderRequest, EmbedderVector};
 use crate::providers::provider::ProviderID;
 use crate::run::Credentials;
+use crate::search_filter::{Filterable, SearchFilter};
 use crate::stores::store::Store;
 use crate::utils;
 use anyhow::{anyhow, Result};
