+++ core/src/lib.rs
@@ -21,9 +21,9 @@ pub mod databases {
     pub mod database;
     pub mod table_schema;
 }
-pub mod filter;
 pub mod project;
 pub mod run;
+pub mod search_filter;
 pub mod utils;
 pub mod providers {
     pub mod azure_openai;
